+++ api/src/main/java/org/allaymc/api/block/interfaces/BlockOxeyeDaisyBehavior.java
@@ -1,6 +1,0 @@
-package org.allaymc.api.block.interfaces;
-
-import org.allaymc.api.block.BlockBehavior;
-
-public interface BlockOxeyeDaisyBehavior extends BlockBehavior {
-}