+++ api/src/main/java/org/allaymc/api/block/interfaces/BlockTorchflowerCropBehavior.java
@@ -1,0 +1,6 @@
+package org.allaymc.api.block.interfaces;
+
+import org.allaymc.api.block.BlockBehavior;
+
+public interface BlockTorchflowerCropBehavior extends BlockBehavior {
+}