+++ api/src/main/java/org/allaymc/api/item/interfaces/ItemDandelionStack.java
@@ -1,6 +1,0 @@
-package org.allaymc.api.item.interfaces;
-
-import org.allaymc.api.item.ItemStack;
-
-public interface ItemDandelionStack extends ItemStack {
-}