+++ api/src/main/java/org/allaymc/api/item/interfaces/ItemTorchflowerSeedsStack.java
@@ -1,0 +1,6 @@
+package org.allaymc.api.item.interfaces;
+
+import org.allaymc.api.item.ItemStack;
+
+public interface ItemTorchflowerSeedsStack extends ItemStack {
+}