+++ server/src/main/java/org/allaymc/server/block/impl/BlockTorchflowerCropBehaviorImpl.java
@@ -1,0 +1,14 @@
+package org.allaymc.server.block.impl;
+
+import org.allaymc.api.block.interfaces.BlockTorchflowerCropBehavior;
+import org.allaymc.api.component.Component;
+import org.allaymc.server.component.ComponentProvider;
+
+import java.util.List;
+
+public class BlockTorchflowerCropBehaviorImpl extends BlockBehaviorImpl implements BlockTorchflowerCropBehavior {
+    public BlockTorchflowerCropBehaviorImpl(
+            List<ComponentProvider<? extends Component>> componentProviders) {
+        super(componentProviders);
+    }
+}